+++ lib/active_record_compose/delegate_attribute.rb
@@ -54,6 +54,7 @@ module ActiveRecordCompose
         end
 
         delegate(*delegates, to:, allow_nil:)
+        attributes.each { define_attribute_methods _1 }
         self.delegated_attributes = delegated_attributes.to_a + attributes.map { _1.to_s }
       end
     end